Split CodePaths and FileCodeSnippet into multiple files
